Crate problem103

Problem 103 (Special subset sums: optimum)

Solution detail

There is a reasonably efficient method of iterating through all special sum sets, in increasing order of their average element. It relies on three observations:

Special sum sets can be generated by adding one element at a time, starting from the sets containing only two elements, and always adding elements in increasing order.

When adding a new element to an existing SSS, checking condition (i) can be done using a merge-sort-like algorithm on two lists - one of the set of sums in the original set, and one of the set of sums of the new set that include the new element.

When adding a new element, condition (ii) is equivalent to ensuring that the added element does not exceed a particular value which is easily calculated from the existing set.

With these observations in mind, the way to generate all special sum sets in increasing order of average element is:

Start with a priority queue (keyed by average element) of special sum sets, initially containing only the set {1, 2}.

Repeatedly pop the minimum element of the queue, and add its children to the queue, where the children are all SSS's that result from adding a single element to the current one.

An SSS of size 2 also has some "special" children - {a, b} --> {a, b + 1} and, if b - a = 1 then also {a, b} --> {a + 1, b + 1}. This ensures all "starting" sets of size 2 are considered.

Because any child of an SSS has a strictly greater average than the original SSS, this ensures that all SSS's are seen in increasing order of average, so to find an optimum SSS containing 7 elements, just stop the iteration when the first 7-element set is found.

Structs

SpecialSumSet
A structure representing a special sum set

SpecialSumSetIterator
A structure capable of iterating over SpecialSumSets using a priority queue.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find an optimum special sum set containing the given number of elements.